move ThreadDump to test package and removed unused methods
